--- src/Files.App/Actions/FileSystem/OpenItemWithApplicationPickerAction.cs
@@ -1,41 +1,0 @@
-﻿using CommunityToolkit.Mvvm.ComponentModel;
-using CommunityToolkit.Mvvm.DependencyInjection;
-using Files.App.Commands;
-using Files.App.Contexts;
-using Files.App.Extensions;
-using Files.App.Helpers;
-using System.ComponentModel;
-using System.Linq;
-using System.Threading.Tasks;
-using Windows.Storage;
-
-namespace Files.App.Actions
-{
-	internal class OpenItemWithApplicationPickerAction : ObservableObject, IAction
-	{
-		private readonly IContentPageContext context = Ioc.Default.GetRequiredService<IContentPageContext>();
-
-		public string Label => "BaseLayoutItemContextFlyoutOpenItemWith/Text".GetLocalizedResource();
-
-		public RichGlyph Glyph => new(opacityStyle: "ColorIconOpenWith");
-
-		public bool IsExecutable => context.HasSelection && context.SelectedItems.All(
-				i => (i.PrimaryItemAttribute == StorageItemTypes.File && !i.IsShortcut && !i.IsExecutable) || (i.PrimaryItemAttribute == StorageItemTypes.Folder && i.IsArchive));
-
-		public OpenItemWithApplicationPickerAction()
-		{
-			context.PropertyChanged += Context_PropertyChanged;
-		}
-
-		public async Task ExecuteAsync()
-		{
-			NavigationHelpers.OpenSelectedItems(context.ShellPage, true);
-		}
-
-		private void Context_PropertyChanged(object? sender, PropertyChangedEventArgs e)
-		{
-			if (e.PropertyName is nameof(IContentPageContext.HasSelection))
-				OnPropertyChanged(nameof(IsExecutable));
-		}
-	}
-}
